+++ src/Test/Scene/wasmRuntimeTestScene.ts
@@ -10,11 +10,10 @@ import { ImageProcessingConfiguration } from "@babylonjs/core/Materials/imagePro
 import { Color4 } from "@babylonjs/core/Maths/math.color";
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
 import { TransformNode } from "@babylonjs/core/Meshes/transformNode";
-import { HavokPlugin } from "@babylonjs/core/Physics/v2/Plugins/havokPlugin";
 import { DepthOfFieldEffectBlurLevel } from "@babylonjs/core/PostProcesses/depthOfFieldEffect";
 import { DefaultRenderingPipeline } from "@babylonjs/core/PostProcesses/RenderPipeline/Pipelines/defaultRenderingPipeline";
 import { Scene } from "@babylonjs/core/scene";
-import havokPhysics from "@babylonjs/havok";
+import ammo from "ammojs-typed";
 
 import type { MmdAnimation } from "@/Loader/Animation/mmdAnimation";
 import type { MmdStandardMaterialBuilder } from "@/Loader/mmdStandardMaterialBuilder";
@@ -27,7 +26,8 @@ import type { MmdMesh } from "@/Runtime/mmdMesh";
 import { MmdWasmInstanceTypeMD } from "@/Runtime/Optimized/InstanceType/multiDebug";
 import { getMmdWasmInstance } from "@/Runtime/Optimized/mmdWasmInstance";
 import { MmdWasmRuntime, MmdWasmRuntimeAnimationEvaluationType } from "@/Runtime/Optimized/mmdWasmRuntime";
-import { MmdPhysics } from "@/Runtime/Physics/mmdPhysics";
+import { MmdAmmoJSPlugin } from "@/Runtime/Physics/mmdAmmoJSPlugin";
+import { MmdAmmoPhysics } from "@/Runtime/Physics/mmdAmmoPhysics";
 import { MmdPlayerControl } from "@/Runtime/Util/mmdPlayerControl";
 
 import type { ISceneBuilder } from "../baseRuntime";
@@ -73,7 +73,7 @@ export class SceneBuilder implements ISceneBuilder {
                 const mmdWasmInstance = await getMmdWasmInstance(new MmdWasmInstanceTypeMD());
                 updateProgress({ lengthComputable: true, loaded: 1, total: 1 });
 
-                const mmdRuntime = new MmdWasmRuntime(mmdWasmInstance, scene, new MmdPhysics(scene));
+                const mmdRuntime = new MmdWasmRuntime(mmdWasmInstance, scene, new MmdAmmoPhysics(scene));
                 mmdRuntime.loggingEnabled = true;
                 mmdRuntime.evaluationType = MmdWasmRuntimeAnimationEvaluationType.Immediate;
 
@@ -116,9 +116,9 @@ export class SceneBuilder implements ISceneBuilder {
             }],
             ["physics", async(updateProgress): Promise<void> => {
                 updateProgress({ lengthComputable: true, loaded: 0, total: 1 });
-                const havokInstance = await havokPhysics();
-                const havokPlugin = new HavokPlugin(true, havokInstance);
-                scene.enablePhysics(new Vector3(0, -9.8 * 10, 0), havokPlugin);
+                const ammoInstance = await ammo();
+                const ammoPlugin = new MmdAmmoJSPlugin(true, ammoInstance);
+                scene.enablePhysics(new Vector3(0, -9.8 * 10, 0), ammoPlugin);
                 updateProgress({ lengthComputable: true, loaded: 1, total: 1 });
             }]
         ]);